Fix: Scripts/CameraCtrl> Problem that camera's x axis rotation became MAX_ANGLE when it's less than 0 degrees
This seems to because of Unity's unexpected work.
Even if localEulerAngle.x is negative value on Inspector view,
It returns in form of '0~360' angle, not '-180~180`.
When we set x angle to -100, localEulerAngle.x returns not '-100' but '260'.
So if I clamp it with range of `-90~90`, the result will be '90'.

Some Japanese infos:       https://twitter.com/Cj_bc_sd/status/1196595991136198656?s=20
And gist for this problem: https://gist.github.com/Cj-bc/1ddfe4a57b18a7f9339aa3d0cff155b7